requirements.txt
openai
python-dotenv
pandas
google-api-python-client
google-auth
google-auth-oauthlib
slack_bolt
flask
